update ArcGIS tool box for OGR

+++ pyfiles/D8ContributingArea.py
@@ -89,8 +89,9 @@ arcpy.AddMessage('Executing: Calculate Statistics\n')
 arcpy.CalculateStatistics_management(ad8)
 
 # remove converted json file
-extn_json=os.path.splitext(shfl)[1] # get extension of the converted json file
-if extn_json==".json":
+if arcpy.Exists(ogrfile):
+  extn_json=os.path.splitext(shfl)[1] # get extension of the converted json file
+  if extn_json==".json":
     os.remove(shfl)
