+++ pyfiles/GridNetwork.py
@@ -87,6 +87,7 @@ arcpy.CalculateStatistics_management(gord)
 arcpy.CalculateStatistics_management(plen)
 arcpy.CalculateStatistics_management(tlen)
 # remove converted json file
-extn_json=os.path.splitext(shfl)[1] # get extension of the converted json file
-if extn_json==".json":
+if arcpy.Exists(ogrfile):
+  extn_json=os.path.splitext(shfl)[1] # get extension of the converted json file
+  if extn_json==".json":
     os.remove(shfl)
